+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Pebble: Allow to select the preferred activity tracker via settings activity (Health, Misfit, Morpheuz)
 * Pebble: Fix wrong(previous) contact being displayed on the pebble 
 * Mi Band: improvements to pairing
+* Fix a problem related to shared preferences storage of activity settings.
 
 ####Version 0.7.4
 * Refactored the settings activity: User details are now generic instead of miband specific. Old settings are preserved.